+++ oshi-core/src/main/java/oshi/driver/linux/Devicetree.java
@@ -1,11 +1,12 @@
 /*
- * Copyright 2020-2022 The OSHI Project Contributors
+ * Copyright 2020-2024 The OSHI Project Contributors
  * SPDX-License-Identifier: MIT
  */
 package oshi.driver.linux;
 
 import oshi.annotation.concurrent.ThreadSafe;
 import oshi.util.FileUtil;
+import oshi.util.platform.linux.SysPath;
 
 /**
  * Utility to read info from the devicetree
@@ -22,7 +23,7 @@ public final class Devicetree {
      * @return The model if available, null otherwise
      */
     public static String queryModel() {
-        String modelStr = FileUtil.getStringFromFile("/sys/firmware/devicetree/base/model");
+        String modelStr = FileUtil.getStringFromFile(SysPath.MODEL);
         if (!modelStr.isEmpty()) {
             return modelStr.replace("Machine: ", "");
         }