+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 * Your contribution here!
 
+##### Bug fixes / Improvements
+* [#2605](https://github.com/oshi/oshi/pull/2605): Reduce CpuStat.getSystemCpuLoadticks memory allocation pressure - [@chrisribble](https://github.com/chrisribble).
+
 # 6.5.0 (2024-03-10)
 
 ##### New Features